task: revert original your-files directory
