+++ dataworkspace/dataworkspace/static/js/react/features/your-files/Header.jsx
@@ -1,18 +1,17 @@
-import React from 'react';
+import React from "react";
+import { TrashIcon } from "./icons/trash";
+import { UploadIcon } from "./icons/upload";
+import { NewFolderIcon } from "./icons/newfolder";
 
-import NewFolderIcon from '../../icons/NewFolderIcon';
-import TrashIcon from '../../icons/TrashIcon';
-import UploadIcon from '../../icons/UploadIcon';
-
-export default class Header extends React.Component {
+export class Header extends React.Component {
   constructor(props) {
     super(props);
-    this.state = { spin: false };
+    this.state = { spin: false }
   }
 
   handleRefreshClick() {
     this.props.onRefreshClick();
-    this.setState({ spin: true });
+    this.setState({ spin: true })
   }
 
   render() {
