--- dataworkspace/dataworkspace/static/js/react/features/your-files/components/PopupTypes/index.jsx
@@ -1,7 +1,0 @@
-const PopupTypes = {
-    ADD_FOLDER: 'addFolder',
-    UPLOAD_FILES: 'uploadFiles',
-    DELETE_OBJECTS: 'deleteObjects',
-};
-
-export default PopupTypes;
