+++ dataworkspace/dataworkspace/static/js/react/features/your-files/icons/upload.js
@@ -1,13 +1,12 @@
-import React from 'react';
+import React from "react";
+import classNames from "classnames";
 
-import classNames from 'classnames';
-
-export default function UploadIcon(props) {
-  const svgClasses = classNames('svg-inline--fa', 'fa-cloud-upload-alt', {
-    'navbutton-icon': props.isNavIcon,
-    'button-icon': !props.isNavIcon,
-    'fa-w-14': !props.isNavIcon,
-    'fa-w-20': props.isNavIcon,
+export function UploadIcon(props) {
+  const svgClasses = classNames("svg-inline--fa", "fa-cloud-upload-alt", {
+    "navbutton-icon": props.isNavIcon,
+    "button-icon": !props.isNavIcon,
+    "fa-w-14": !props.isNavIcon,
+    "fa-w-20": props.isNavIcon,
   });
 
   return (
@@ -20,8 +19,8 @@ export default function UploadIcon(props) {
       role="img"
       xmlns="http://www.w3.org/2000/svg"
       viewBox="0 0 640 512"
-      width={props.isNavIcon ? '60' : null}
-      height={props.isNavIcon ? '60' : null}
+      width={props.isNavIcon ? "60" : null}
+      height={props.isNavIcon ? "60" : null}
     >
       <path
         fill="currentColor"
