+++ dataworkspace/dataworkspace/static/js/react/features/your-files/popups/Modal.jsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useRef } from 'react';
-
 import styled from 'styled-components';
 
 const Dialog = styled('dialog')`
